Deleted folder "News" and moved all the tests to the folder "Newss"
